==> part1.ipynb <==
# Import Libraries

import os
import glob
import pandas as pd
import numpy as np

# Download

basics_url = 'https://datasets.imdbws.com/title.basics.tsv.gz'
akas_url = 'https://datasets.imdbws.com/title.akas.tsv.gz'
ratings_url = 'https://datasets.imdbws.com/title.ratings.tsv.gz'

# Title Basics

basics = pd.read_csv(basics_url, sep = '\t', low_memory = False)

basics.head()

## Process Data

### Replace null values (\N) with NAN


basics.replace({'\\N':np.nan}, inplace = True)

basics.head()

### Eliminate movies that are null for 'runtimeMinutes' and 'genres'

basics = basics.dropna(subset = ['runtimeMinutes', 'genres'])

### Keep only titleType==Movie

basics = basics.loc[basics['titleType'] == 'movie']
basics.head()

### Keep startYear 2000-2022

basics.info()

The type of 'startYear' is object.

basics = basics.dropna(subset = ['startYear'])
basics['startYear'] = basics['startYear'].astype('int64')
basics = basics.loc[(basics['startYear']>= 2000) & (basics['startYear']<= 2022)]
basics.head()

### Eliminate movies that include "Documentary" in genre

is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]
basics.head()

## Make the Data folder if it doesn't already exist
os.makedirs('Data',exist_ok=True)

# Confirm folder created
os.listdir("Data/")

## Save current dataframe to file.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)

# Open saved file and preview again
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics.head()

basics.info()

# Title Ratings

ratings = pd.read_csv(ratings_url, sep = '\t', low_memory = False)

ratings.head()

## Replace "\N" with np.nan (if any)

ratings.replace({'\\N':np.nan}, inplace = True)

## Save current dataframe to file.
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

# Open saved file and preview again
ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings.head()

ratings.info()

# Title Akas

chunk_num = 1
akas_reader = pd.read_csv(akas_url, sep='\t', low_memory=False, chunksize=100_000)

for temp_akas in akas_reader:
        #### COMBINED WORKFLOW FROM ABOVE
    ## Replace "\N" with np.nan
    temp_akas.replace({'\\N':np.nan},inplace=True)
    
    ## keep only US movies
    temp_akas = temp_akas.loc[temp_akas['region']=='US']

    ### Saving chunk to disk
    fname= f'Data/title_akas_chunk_{chunk_num:03d}.csv.gz'
    temp_akas.to_csv(fname, compression='gzip')
    print(f"- Saved {fname}")
    
    ## increment chunk_num    
    chunk_num+=1
## Closing the reader now that we are done looping through the file
akas_reader.close()

Tip: if we use the ":03d" format code when inserting the chunk number using an f-string, it will add 2 leading 0's, so the first file will be numbered 001 instead of 1. This will be helpful when viewing the files in your file explorer or on GitHub.

q = "Data/title_akas_chunk*.csv.gz"
chunked_files = glob.glob(q)
# Showing the first 5
chunked_files[:5]

## Loading all files as df and appending to a list
akas_list = []
for file in chunked_files:
    temp_akas = pd.read_csv(file, index_col=0)
    akas_list.append(temp_akas)
    
## Concatenating the list of dfs into 1 combined
akas_combined = pd.concat(akas_list)
akas_combined.head()

## Saving the final combined dataframe
final_fname ='Data/title_akas_combined.csv.gz'
akas_combined.to_csv(final_fname, compression='gzip', index=False)

# Open saved file and preview again
akas_combined = pd.read_csv(final_fname)
akas_combined.head()

akas_combined.info()

# Filtering one dataframe based on another

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers_basics =basics['tconst'].isin(akas_combined['titleId'])
keepers_ratings = ratings['tconst'].isin(akas_combined['titleId'])

basics = basics[keepers_basics]
basics

ratings = ratings[keepers_ratings]
ratings

## Save current dataframe to file.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)
