==> part2.1.ipynb <==
# Source of Data
The Movie Database (TMDB): https://www.themoviedb.org/

This notebook is created to extract tmdb data fromm 2002 to 2022 

# Import Libraries

import pandas as pd
import os, time,json
import tmdbsimple as tmdb 
from tqdm.notebook import tqdm_notebook
import seaborn as sns
import matplotlib.pyplot as plt

# API Credentials

with open('C://Users//sheha//.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
    
## Display the keys of the loaded dict
login.keys()

tmdb.API_KEY =  login['api-key']

# Designate a folder

FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok= True)
os.listdir(FOLDER)

# Define Functions

## Movie Rating Function

def get_movie_with_rating(movie_id):
    """Adapted  from source =  https://github.com/celiao/tmdbsimple"""
    # Get the object for the current id
    movie = tmdb.Movies(movie_id)
    
    # save the .info .releases dictionaries 
    info = movie.info()
    
    releases = movie.releases()
    # loop through countries in releases
    for c in releases['countries']:
        if c['iso_3166_1'] == 'US':
            ## save a "certification" key in the info dict with the certification
            info['certification'] = c['certification']
    return info

## Append records to json file

def write_json(new_data, filename): 
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
    
    with open(filename, 'r+') as file:
        #First we load existing data into a dict.
        file_data = json.load(file)
        ## use extend and or append
        if (type(new_data)==list) & (type(file_data)==list):
            file_data.extend(new_data)
        else:
            file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        #convert back to json
        json.dump(file_data, file)

# Load in the Title Basics data

basics = pd.read_csv("Data/title_basics.csv.gz")

# Create Required Lists for the Loop

# Define a list of the Years to Extract from the API
YEARS_TO_GET = [*range(2002, 2023,1)]
YEARS_TO_GET 

# define error list
errors = [ ]

# Testing the "get_movie_with_rating" function

test_ids = ["tt0848228", "tt0332280"]
results = []

for movie_id in test_ids:
    
    try:
        movie_info = get_movie_with_rating(movie_id)
        results.append(movie_info)
        
    except Exception as e:
        errors.append([movie_id, e])
    
    
pd.DataFrame(results)

print(f"- Number of errors: {len(errors)}")
errors

# Set up Progress Bar

# OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    #check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    # If it does not exist: create it
    if file_exists == False:
        # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
    
    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    
    # saving movie ids to list
    movie_ids = df['tconst'].copy()
    
    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]

# INNER Loop
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        except Exception as e:
            errors.append([movie_id, e])

    
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz",
                         compression="gzip", index=False)



tmdb_data_all = pd.read_csv("Data/tmdb_results_combined.csv.gz")
tmdb_data_all = tmdb_data_all.loc[1:]
tmdb_data_all.head(2)

len(tmdb_data_all)

for i in YEARS_TO_GET:
    df = pd.read_csv(f"Data/final_tmdb_data_{i}.csv.gz")
    df = df.loc[1:]
    tmdb_data_all = pd.concat([tmdb_data_all, df], ignore_index= True)
    print(i, len(df) ,len(tmdb_data_all))

display (tmdb_data_all.head(2), tmdb_data_all.tail(2))

# Save a final merged .csv.gz of all of the tmdb api data 
tmdb_data_all.to_csv("Data/tmdb_data_all.csv.gz", 
                                compression="gzip", 
                                index=False)
