==> part3.ipynb <==
# Part3: Construct and export a MySQL database

# Import Libraries

import pandas as pd
import os, time,json
import tmdbsimple as tmdb 
from tqdm.notebook import tqdm_notebook
import seaborn as sns
import matplotlib.pyplot as plt

import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
from sqlalchemy.types import *

# Normalizing Genres

# load title_basics csv file
title_basics = pd.read_csv("Data/title_basics.csv.gz")
title_basics.head()

## Getting List of Unique Genres

test = title_basics['genres'][0]
type(test)

# create a new column with the single string genres as a list of strings
title_basics['genres_split'] =title_basics['genres'].str.split(',')
title_basics.head()

test = title_basics['genres_split'][0]
type(test)

# Separate the list of genres into new rows 
exploded_genres = title_basics.explode('genres_split')
exploded_genres

# get the unique genres from the genres_split column
unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres

## Create a new title_genres table

# create a new title_genres from the exploded_genres
title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head()

## Create a genre mapper dictionary to replace string genres with integers

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## Replace the string genres in title_genres with the new integer ids

title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres

# Drop the "genres_split" column
title_genres.drop(columns = ['genres_split'], axis = 1, inplace = True)
title_genres

## Convert the genre map dictionary into a dataframe

genres = pd.DataFrame({'genre_name': genre_map.keys(),
                            'genre_id': genre_map.values()})
genres.head()

# Discard unnecessary information in tables

title_basics.head()

title_basics.info()

#drop unneecessary cilumns from titile_basics table
title_basics.drop(columns = ["originalTitle", "isAdult", "titleType", "genres", "genres_split"], axis =1, inplace = True)
title_basics.info()

# Create SQL Movies Database

# Creating the connection 
username = "root"
password = "root" 
db_name = "Movies"
connection = f"mysql+pymysql://{username}:{password}@localhost/{db_name}"

#creating the engine 
engine = create_engine(connection)
engine

#creatiing and checking the existence of new database
if database_exists(connection) == True:
    print('The database is already exist')
else:
    create_database(connection)

## title_basics table

### Creating a dataype schema for to_sql

## Calculate max string lengths for object columns
key_len = title_basics['tconst'].fillna('').map(len).max()
title_len = title_basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Integer(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

### Run table.to_sql with the dtype argument

# Save to sql with dtype and index=False
title_basics.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)

### Run the query to ADD PRIMARY KEY

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

# show the first 5 rows
q = """
select *
from title_basics LIMIT 5;
"""
pd.read_sql(q, engine)

## title_ratings table

# load ratings csv file
title_ratings = pd.read_csv("Data/title_ratings.csv.gz")
title_ratings.head()

title_ratings.info()

### Creating a dataype schema for to_sql

## Calculate max string lengths for object columns
key_len = title_ratings['tconst'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    'averageRating': Float(),
    'numVotes':Integer()}

### Run table.to_sql with the dtype argument

# Save to sql with dtype and index=False
title_ratings.to_sql('title_ratings',engine,dtype=df_schema,if_exists='replace',index=False)

### Run the query to ADD PRIMARY KEY

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

# show the first 5 rows
q = """
select *
from title_ratings LIMIT 5;
"""
pd.read_sql(q, engine)

## title_genres table

title_genres.info()

### Creating a dataype schema for to_sql

## Calculate max string lengths for object columns
tconst_len = title_genres['tconst'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": Text(tconst_len+1), 
    'genre_id':Integer()}

### Run table.to_sql with the dtype argument

# Save to sql with dtype and index=False
title_genres.to_sql('title_genres',engine, dtype=df_schema, if_exists='replace',index=False)

# show the first 5 rows
q = """
select *
from title_genres LIMIT 5;
"""
pd.read_sql(q, engine)

## genres table


title_genres.info()

### Creating a dataype schema for to_sql

## Calculate max string lengths for object columns
genrename_len = genres['genre_name'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "genre_name": String(genrename_len+1), 
    'genre_id':Integer()}

### Run table.to_sql with the dtype argument

# Save to sql with dtype and index=False
genres.to_sql('genres',engine,dtype=df_schema, if_exists='replace',index=False)

### Run the query to ADD PRIMARY KEY

engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`genre_id`);')

# show the first 5 rows
q = """
select *
from genres LIMIT 5;
"""
pd.read_sql(q, engine)

## tmdb_data table

# load csv file
tmdb_data = pd.read_csv("Data/tmdb_results_combined.csv.gz")
tmdb_data.head()

tmdb_data.info()

tmdb_data = tmdb_data[['imdb_id','revenue','budget','certification']]
tmdb_data = tmdb_data.iloc[1:]
tmdb_data

tmdb_data.info()

### Creating a dataype schema for to_sql

## Calculate max string lengths for object columns
id_len = tmdb_data['imdb_id'].fillna('').map(len).max()
certi_len = tmdb_data['certification'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "imdb_id": String(id_len+1), 
    'revenue': Float(),
    'budget': Float(),
    "certification": String(certi_len+1)}

### Run table.to_sql with the dtype argument

# Save to sql with dtype and index=False
tmdb_data.to_sql('tmdb_data',engine, dtype=df_schema, if_exists='replace', index=False)

### Run the query to ADD PRIMARY KEY

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

# show the first 5 rows
q = """
select *
from tmdb_data LIMIT 5;
"""
pd.read_sql(q, engine)

## Show tables

# Display the SQL table names using an SQL query
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
