==> part5.ipynb <==
# Part5: Produce a Linear Regression model to predict movie performance

# Import Libraries

import os
import glob
import json
import pandas as pd
import numpy as np
import seaborn as sns
from scipy import stats
import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.impute import SimpleImputer
from sklearn.compose import make_column_transformer, make_column_selector, ColumnTransformer
from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.metrics import r2_score,mean_absolute_error,mean_squared_error
## fixing random for lesson generation
np.random.seed(321)

# Load Data

FOLDER = 'Data/'
file_list = sorted(os.listdir(FOLDER))
file_list

#listing required files
path = 'Data/final_*.csv.gz'
file_list = sorted(glob.glob(path))
file_list

#combining all files into one dataframe
df = pd.concat(pd.read_csv(f,lineterminator='\n') for f in file_list)
df.shape

df.head(2)

# Data Cleaning

#remove id = 0
df = df.loc[df['imdb_id']!='0']
df.head(2)

#drop unwanted columns
drop_cols = ['backdrop_path','backdrop_path','original_title','overview',
                 'poster_path','status','tagline','id','homepage',
                 'production_countries','video','production_companies','spoken_languages',
            'original_language']
df.drop(columns = drop_cols, inplace = True, axis =1)
df.shape

#converting 'belongs_to_collection' into boolean
df['belongs_to_collection'] =df['belongs_to_collection'].notna()
df['belongs_to_collection'].value_counts()
